requirements.txt
pymlask==0.2.3
mecab-python3==0.996.5
